requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
google-genai>=0.3.0
google-generativeai>=0.8.0
websockets>=12.0
tenacity>=8.0.0
typing-extensions>=4.5.0
httpx>=0.25.0